requirements.txt
# Core dependencies
langgraph>=0.2.0,<0.3.0
langchain>=0.3.0,<0.4.0
langchain-groq>=0.2.0,<0.3.0
python-dotenv>=1.0.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0

# Utilities
dataclasses-json>=0.6.0
colorama>=0.4.6
tabulate>=0.9.0